ml_dsa

Module pkcs8

PKCS#8 private key encoding support.

Modules

der

RustCrypto: ASN.1 DER

spki

RustCrypto: X.509 Subject Public Key Info (SPKI)

Structs

Document

ASN.1 DER-encoded document.

ObjectIdentifier

Object identifier (OID).

PrivateKeyInfo

PKCS#8 PrivateKeyInfo.

SecretDocument

Secret Document type.

SubjectPublicKeyInfo

X.509 SubjectPublicKeyInfo (SPKI) as defined in RFC 5280  4.1.2.7.

Enums

Error

Error type

KeyError

Key-related errors.

LineEnding

Line endings: variants of newline characters that can be used with Base64.

Version

Version identifier for PKCS#8 documents.

Constants

SEED_TAG_NUMBER 🔒

Tag number for the seed value.

Traits

AssociatedOid

A trait which associates an OID with a type.

DecodePrivateKey

Parse a private key object from a PKCS#8 encoded document.

DecodePublicKey

Parse a public key object from an encoded SPKI document.

EncodePrivateKey

Serialize a private key object to a PKCS#8 encoded document.

EncodePublicKey

Serialize a public key object to a SPKI-encoded document.

Type Aliases

AlgorithmIdentifierRef

AlgorithmIdentifier reference which has AnyRef parameters.

PrivateKeyInfoOwned

PrivateKeyInfo with Any algorithm parameters, and Box<[u8]> key.

PrivateKeyInfoRef

PrivateKeyInfo with AnyRef algorithm parameters, and &[u8] key.

Result

Result type

SeedString 🔒

ML-KEM seed serialized as ASN.1.

SubjectPublicKeyInfoRef

SubjectPublicKeyInfo with AnyRef algorithm parameters, and BitStringRef params.