ml_dsa::pkcs8::der::oid

Type Alias Arc

pub type Arc = u32;

Type alias used to represent an “arc”, i.e. integer identifier value, where an OID comprises a sequence of arcs.

X.660 does not define a maximum size of an arc. We instead follow Mozilla* conventions for maximum values of an arc, with a maximum value of 2^32-1 (4294967295), a.k.a. u32::MAX with Arc being a type alias for u32.

Note that this means we deliberately do NOT support UUIDs used as OIDs.

*NOTE: please see this study for a survey of how various OID libraries handle maximum arcs: https://misc.daniel-marschall.de/asn.1/oid_facts.html