ml_dsa::pkcs8::der

Struct SliceReader

pub struct SliceReader<'a> {
    bytes: &'a BytesRef,
    encoding_rules: EncodingRules,
    failed: bool,
    position: Position,
}

Reader which consumes an input byte slice.

Fields
bytes: &'a BytesRefencoding_rules: EncodingRulesfailed: boolposition: Position
Implementations

impl<'a> SliceReader<'a>

pub fn new(bytes: &'a [u8]) -> Result<SliceReader<'a>, Error>

Create a new slice reader for the given byte slice.

Errors

If bytes is too long.

pub fn new_with_encoding_rules( bytes: &'a [u8], encoding_rules: EncodingRules, ) -> Result<SliceReader<'a>, Error>

Create a new slice reader with the given encoding rules.

Errors

If bytes is too long.

pub fn error(&mut self, kind: ErrorKind) -> Error

Return an error with the given ErrorKind, annotating it with context about where the error occurred.

pub fn is_failed(&self) -> bool

Did the decoding operation fail due to an error?