ml_dsa::signing

Struct SigningKey

pub struct SigningKey<P: MlDsaParams> {
    expanded_key: MaybeBox<ExpandedSigningKey<P>>,
    seed: MaybeBox<Seed>,
    verifying_key: VerifyingKey<P>,
}

ML-DSA signing key (i.e. private/secret key).

This type is initialized through a Seed, and can be used to generate ML-DSA signatures.

Fields
expanded_key: MaybeBox<ExpandedSigningKey<P>>

The expanded form of the signing key.

seed: MaybeBox<Seed>

The seed this signing key was derived from

verifying_key: VerifyingKey<P>

When the alloc feature is available, precompute the VerifyingKey.

Implementations

impl<P: MlDsaParams> SigningKey<P>

pub fn from_seed(xi: &Seed) -> Self

Deterministically generate a signing key pair from the specified Seed.

This method reflects the ML-DSA.KeyGen_internal algorithm from FIPS 204 (Algorithm 6).

pub fn as_seed(&self) -> &Seed

Borrow the Seed value: 32-bytes which can be used to reconstruct the SigningKey.

Warning
This value is key material. Please treat it with care.

pub fn to_seed(&self) -> Seed

Serialize the Seed value: 32-bytes which can be used to reconstruct the SigningKey.

Warning
This value is key material. Please treat it with care.