ml_dsa

Struct VerifyingKey

pub struct VerifyingKey<P: ParameterSet> {
    rho: B32,
    t1: MaybeBox<Vector<BaseField, P::K>>,
    precomputed_values: MaybeBox<PrecomputedValues<P>>,
}

An ML-DSA verification key.

Fields
rho: B32

Public seed used to deterministically re-expand A_hat.

t1: MaybeBox<Vector<BaseField, P::K>>

High bits of the public key polynomial t.

precomputed_values: MaybeBox<PrecomputedValues<P>>

Precomputed expanded values.

Implementations

impl<P: MlDsaParams> VerifyingKey<P>

pub(crate) fn new( rho: B32, t1: Vector<BaseField, P::K>, A_hat: NttMatrix<BaseField, P::K, P::L>, enc: Option<EncodedVerifyingKey<P>>, ) -> Self

fn new_expand_a( rho: B32, t1: Vector<BaseField, P::K>, enc: Option<EncodedVerifyingKey<P>>, ) -> Self

pub fn compute_mu<F: FnOnce(&mut Shake256) -> Result<(), Error>>( &self, Mp: F, ctx: &[u8], ) -> Result<Array<u8, U64>, Error>

Computes µ according to FIPS 204 for use in ML-DSA.Sign and ML-DSA.Verify.

Errors

Returns Error if the given Mp returns one.

pub fn verify_internal(&self, M: &[u8], sigma: &Signature<P>) -> bool
where P: MlDsaParams,

Implementation of Algorithm 8: ML-DSA.Verify_internal algorithm from FIPS 204.

It does not include the domain separator that distinguishes between the normal and pre-hashed cases, and it does not separate the context string from the rest of the message.

pub(crate) fn raw_verify_mu( &self, mu: &Array<u8, U64>, sigma: &Signature<P>, ) -> bool
where P: MlDsaParams,

pub fn verify_with_context( &self, M: &[u8], ctx: &[u8], sigma: &Signature<P>, ) -> bool

Implementation of Algorithm 3: ML-DSA.Verify from FIPS 204.

pub fn verify_mu(&self, mu: &Array<u8, U64>, sigma: &Signature<P>) -> bool

Implementation of Algorithm 3: ML-DSA.Verify from FIPS 204 with a pre-computed μ.

fn raw_verify_with_context( &self, M: &[&[u8]], ctx: &[u8], sigma: &Signature<P>, ) -> bool

pub(crate) fn encode_internal( rho: &B32, t1: &Vector<BaseField, P::K>, ) -> EncodedVerifyingKey<P>

pub fn encode(&self) -> EncodedVerifyingKey<P>

Encode the key in a fixed-size byte array.

Implementation of Algorithm 22: pkEncode from FIPS 204.

pub fn decode(enc: &EncodedVerifyingKey<P>) -> Self

Decode the key from an appropriately sized byte array.

Implementation of Algorithm 23: pkDecode from FIPS 204.