ml_dsa

Trait KeyGen

pub trait KeyGen: MlDsaParams {
    type KeyPair: Keypair;

    // Required methods
    fn key_gen<R: CryptoRng + ?Sized>(rng: &mut R) -> Self::KeyPair;
    fn from_seed(xi: &B32) -> Self::KeyPair;
}

A parameter set that knows how to generate key pairs

Required Associated Types

type KeyPair: Keypair

The type that is returned by key generation

Required Methods

fn key_gen<R: CryptoRng + ?Sized>(rng: &mut R) -> Self::KeyPair

Generate a signing key pair from the specified RNG

fn from_seed(xi: &B32) -> Self::KeyPair

Deterministically generate a signing key pair from the specified seed

This method reflects the ML-DSA.KeyGen_internal algorithm from FIPS 204.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.