ml_dsa

Trait MlDsaParams

pub trait MlDsaParams:
    SigningKeyParams
    + VerifyingKeyParams
    + SignatureParams
    + Debug
    + Default
    + PartialEq
    + Clone { }

An instance of MlDsaParams defines all of the parameters necessary for ML-DSA operations. Typically this is done by implementing ParameterSet with values that will fit into the blanket implementations of SigningKeyParams, VerifyingKeyParams, and SignatureParams.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.