Crate ml_dsa

RustCrypto: ML-DSA

Pure Rust implementation of the Module-Lattice-Based Digital Signature Standard (ML-DSA) as described in the FIPS 204 (final).

About

ML-DSA was formerly known as CRYSTALS-Dilithium.

⚠️ Security Warning

The implementation contained in this crate has never been independently audited!

USE AT YOUR OWN RISK!

License

All crates licensed under either of

Apache License, Version 2.0

MIT license

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Quickstart

use ml_dsa::{
    signature::{Keypair, Signer, Verifier},
    MlDsa65, KeyGen,
};
use getrandom::{SysRng, rand_core::UnwrapErr};

let mut rng = UnwrapErr(SysRng);
let kp = MlDsa65::key_gen(&mut rng);

let msg = b"Hello world";
let sig = kp.signing_key().sign(msg);

assert!(kp.verifying_key().verify(msg, &sig).is_ok());

Re-exports

pub use signature;

Structs

Error

Signature errors.

KeyPair

An ML-DSA key pair

MlDsa44

MlDsa44 is the parameter set for security category 2.

MlDsa65

MlDsa65 is the parameter set for security category 3.

MlDsa87

MlDsa87 is the parameter set for security category 5.

Signature

An ML-DSA signature

SigningKey

An ML-DSA signing key

VerifyingKey

An ML-DSA verification key

Traits

KeyGen

A parameter set that knows how to generate key pairs

MlDsaParams

An instance of MlDsaParams defines all of the parameters necessary for ML-DSA operations. Typically this is done by implementing ParameterSet with values that will fit into the blanket implementations of SigningKeyParams, VerifyingKeyParams, and SignatureParams.

Type Aliases

B32

A 32-byte array, defined here for brevity because it is used several times

EncodedSignature

A signature encoded as a byte array

EncodedVerifyingKey

A verifying key encoded as a byte array

ExpandedSigningKey

A signing key encoded as a byte array

Seed

ML-DSA seeds are signing (private) keys, which are consistently 32-bytes across all security levels, and are the preferred serialization for representing such keys.