ml_dsa

Struct SigningKey

pub struct SigningKey<P: MlDsaParams> {
    pub(crate) rho: B32,
    pub(crate) K: B32,
    pub(crate) tr: Array<u8, U64>,
    pub(crate) s1: Vector<BaseField, P::L>,
    pub(crate) s2: Vector<BaseField, P::K>,
    pub(crate) t0: Vector<BaseField, P::K>,
    pub(crate) s1_hat: NttVector<BaseField, P::L>,
    pub(crate) s2_hat: NttVector<BaseField, P::K>,
    pub(crate) t0_hat: NttVector<BaseField, P::K>,
    pub(crate) A_hat: NttMatrix<BaseField, P::K, P::L>,
}

An ML-DSA signing key

Fields
rho: B32K: B32tr: Array<u8, U64>s1: Vector<BaseField, P::L>s2: Vector<BaseField, P::K>t0: Vector<BaseField, P::K>s1_hat: NttVector<BaseField, P::L>s2_hat: NttVector<BaseField, P::K>t0_hat: NttVector<BaseField, P::K>A_hat: NttMatrix<BaseField, P::K, P::L>
Implementations

impl<P: MlDsaParams> SigningKey<P>

pub(crate) fn new( rho: B32, K: B32, tr: Array<u8, U64>, s1: Vector<BaseField, P::L>, s2: Vector<BaseField, P::K>, t0: Vector<BaseField, P::K>, A_hat: Option<NttMatrix<BaseField, P::K, P::L>>, ) -> Self

pub fn sign_internal(&self, Mp: &[&[u8]], rnd: &B32) -> Signature<P>
where P: MlDsaParams,

This method reflects the ML-DSA.Sign_internal algorithm from FIPS 204. It does not include the domain separator that distinguishes between the normal and pre-hashed cases, and it does not separate the context string from the rest of the message.

pub fn sign_randomized<R: RngCore + CryptoRng + ?Sized>( &self, M: &[u8], ctx: &[u8], rng: &mut R, ) -> Result<Signature<P>, Error>

This method reflects the randomized ML-DSA.Sign algorithm.

Errors

This method will return an opaque error if the context string is more than 255 bytes long, or if it fails to get enough randomness.

pub fn sign_deterministic( &self, M: &[u8], ctx: &[u8], ) -> Result<Signature<P>, Error>

This method reflects the optional deterministic variant of the ML-DSA.Sign algorithm.

Errors

This method will return an opaque error if the context string is more than 255 bytes long.

pub fn encode(&self) -> EncodedSigningKey<P>
where P: MlDsaParams,

Encode the key in a fixed-size byte array.

pub fn decode(enc: &EncodedSigningKey<P>) -> Self
where P: MlDsaParams,

Decode the key from an appropriately sized byte array.