ml_dsa

Struct VerifyingKey

pub struct VerifyingKey<P: ParameterSet> { /* private fields */ }

An ML-DSA verification key

Implementations

impl<P: MlDsaParams> VerifyingKey<P>

pub fn compute_mu<F: FnOnce(&mut Shake256) -> Result<(), Error>>( &self, Mp: F, ctx: &[u8], ) -> Result<Array<u8, U64>, Error>

Computes ยต according to FIPS 204 for use in ML-DSA.Sign and ML-DSA.Verify.

Errors

Returns Error if the given Mp returns one.

pub fn verify_internal(&self, M: &[u8], sigma: &Signature<P>) -> bool
where P: MlDsaParams,

This algorithm reflects the ML-DSA.Verify_internal algorithm from FIPS 204. It does not include the domain separator that distinguishes between the normal and pre-hashed cases, and it does not separate the context string from the rest of the message.

pub fn verify_with_context( &self, M: &[u8], ctx: &[u8], sigma: &Signature<P>, ) -> bool

This algorithm reflects the ML-DSA.Verify algorithm from FIPS 204.

pub fn verify_mu(&self, mu: &Array<u8, U64>, sigma: &Signature<P>) -> bool

This algorithm reflects the ML-DSA.Verify algorithm with a pre-computed ฮผ from FIPS 204.

pub fn encode(&self) -> EncodedVerifyingKey<P>

Encode the key in a fixed-size byte array.

pub fn decode(enc: &EncodedVerifyingKey<P>) -> Self

Decode the key from an appropriately sized byte array.